🚀 Added message feature and fixed events

+++ backend/src/chat/chat.service.ts
@@ -8,6 +8,9 @@ import { JwtService } from '@nestjs/jwt';
 import { ro } from '@faker-js/faker';
 import { RoomDto } from './dto/room-dto';
 import { get } from 'http';
+import { Socket } from 'socket.io';
+import { WsException } from '@nestjs/websockets';
+import { AuthWithWs } from './dto/user-ws-dto';
 
 @Injectable()
 export class ChatService {
@@ -16,18 +19,14 @@ export class ChatService {
 
 
 
-  async createDmRoom(socket: any, payload: CreateChatDmRoomDto) {
+  async createRoom(socket: any, payload: CreateChatDmRoomDto) {
     const roomId = await this.generateUniqueRoomId();
-    const role = (payload.role !== 'MEMBER') ? 'OWNER' : 'MEMBER';
+    const role = (payload.type !== 'DM') ? 'OWNER' : 'MEMBER';
 
     if (payload.type === 'PROTECTED' && payload.password === undefined) {
-      throw new Error(`Password is required for protected room`);
+      throw new WsException(`Room password is required`);
     }
-    
-    // here the token should be the user id after the verification
-    console.log("socket id", socket.id);
-    const userID = +socket.id;
-    console.log("roomType : ", payload.type);
+
     const room = await this.prisma.chatRoom.create({
       data: {
         name: roomId,
@@ -35,36 +34,36 @@ export class ChatService {
         password: payload.password,
         members: {
           create: [{
-            user : { connect : { id : userID } },
+            user : { connect : { id : +socket.id } },
             role: role as ChatRole,
           }]
         }
       }
     })
-
     room.password = undefined;
     return room;
   }
 
   async joinRoom(socket: any, payload: any){
+
     const room = await this.prisma.chatRoom.findUnique({
       where: {
         name : payload.name,
       },
       select: {
+        id: true,
         password: true,
         roomType: true,
       }
     });
-    if (!room) {
-      Logger.error(`Room with name ${payload.name} not found`);
-      throw new Error(`Room with name ${payload.name} not found`);
-    }
-    console.log("room : ", room);
-    if (room.roomType === 'PROTECTED' && room.password !== payload.password) {
-      Logger.error(`Password is incorrect`);
-      throw new Error(`Password is incorrect`);
-    }
+    if (!room)
+      throw new WsException(`Room with name ${payload.name} not found`);
+
+    if (room.roomType === 'PROTECTED' && room.password !== payload.password)
+      throw new WsException(`Wrong password`);
+
+    if (await this.isUserInRoom(+room.id, +socket.id) === true)
+      throw new WsException(`You are already a member of this room`);
 
     const patchedRoom = await this.prisma.chatRoom.update({
       where: {
@@ -99,15 +98,31 @@ export class ChatService {
   //   })
   //   return room;
   // }
-
-
   // switch to unique name
   async getRoomByNames(roomName: string): Promise<ChatRoom> {
-    return this.prisma.chatRoom.findUnique({
+    const room = await this.prisma.chatRoom.findUnique({
       where: {
         name: roomName,
+      },
+      select: {
+        id: true,
+        name: true,
+        password: true,
+        roomType: true,
+        updatedAt: true,
+        createdAt: true,
       }
     })
+
+    if  (!room) {
+      throw new WsException(`Room with name ${roomName} not found`);
+    }
+
+    if (room.roomType === 'PRIVATE' || room.roomType === 'DM')
+      throw new WsException(`Room With name ${roomName} doesn't exist or Not Public`);
+    
+
+    return room;
     // hna khas ncheck protect to del and privet to check passorwd
   }
 
@@ -153,7 +168,7 @@ export class ChatService {
   }
 
   
-  async getRoomUsers(roonName: string, params: { // params : { skip?: number; take?: number; cursor?: Prisma.ChatRoomWhereUniqueInput; where?: Prisma.ChatRoomWhereInput; orderBy?: Prisma.ChatRoomOrderByWithRelationInput; }
+  async getRoomUsers(client : AuthWithWs, roonName: string, params: { // params : { skip?: number; take?: number; cursor?: Prisma.ChatRoomWhereUniqueInput; where?: Prisma.ChatRoomWhereInput; orderBy?: Prisma.ChatRoomOrderByWithRelationInput; }
     skip?: number;
     take?: number;
     cursor?: Prisma.ChatRoomWhereUniqueInput;
@@ -164,11 +179,17 @@ export class ChatService {
     const roomId = await this.prisma.chatRoom.findUnique({
       where: {
         name: roonName,
-      }
+      },
     });
     if (!roomId) {
-      throw new Error(`Room with name ${roonName} not found`);
+      throw new WsException(`Room with name ${roonName} not found`);
     }
+
+    if (await this.isUserInRoom(+roomId.id, +client.id) === false) {
+      throw new WsException(`You are not a member of this room`);
+    }
+
+
     const { cursor, orderBy, skip, take} = params;
     const users =  this.prisma.roomMembership.findMany({
       where: {
@@ -200,9 +221,8 @@ export class ChatService {
 
 
   // idea : insetead of delete the user from the room we can just change the state to kicked  
-  async kickUserFromRoom(client: any, payload: any): Promise<void> {
+  async kickUserFromRoom(client: any, payload: any) {
 
-      console.log("name : ", payload.name);
       const room = await this.prisma.chatRoom.findUnique({
         where: {
           name : payload.name,
@@ -216,11 +236,42 @@ export class ChatService {
         }
       });
 
-      await this.prisma.roomMembership.delete({
+      if (await this.isUserAdminInRoom(room.members[0].roomId, +client.id) === false) {
+        throw new WsException(`Not Allowed`);
+      }
+
+      const user = await this.prisma.roomMembership.delete({
         where: {
           id : room.members[0].id,
+        },
+        include: {
+          user: {
+            select: {
+              profile: {
+                select: {
+                  username: true,
+                }
+              }
+            }
+          }
         }
       })
+
+      console.log('user name', user.user.profile.username);
+
+      const kickeduser = await this.prisma.user.findUnique({
+        where: {
+          id: payload.userId,
+        },
+        select: {
+          profile: {
+            select: {
+              username: true,
+            }
+          }
+        }
+      })
+      return kickeduser;
   }
   
 
@@ -267,12 +318,45 @@ export class ChatService {
         }
       }
     });
+
+    if (await this.isUserAdminInRoom(room.members[0].roomId, +client.id) === false) {
+      throw new WsException(`Not Allowed`);
+    }
+
     await this.prisma.roomMembership.update({
       where: {
         id : room.members[0].id,
       },
       data: {
         state: 'BANNED',
+      }
+    })
+  }
+
+  async UnBanUserFromRoom(client: any, payload: any): Promise<void> {
+    const room = await this.prisma.chatRoom.findUnique({
+      where: {
+        name : payload.name,
+      },
+      select: {
+        members: {
+          where: {
+            userId: payload.userId,
+          }
+        }
+      }
+    });
+
+    if (await this.isUserAdminInRoom(room.members[0].roomId, +client.id) === false) {
+      throw new WsException(`Not Allowed`);
+    }
+
+    await this.prisma.roomMembership.update({
+      where: {
+        id : room.members[0].id,
+      },
+      data: {
+        state: 'ACTIVE',
       }
     })
   }
@@ -301,6 +385,11 @@ export class ChatService {
       return;
     }
 
+    if (room.members[0].state === 'BANNED') {
+      client.emit('roomBroadcast', "You are banned");
+      return;
+    }
+
     const message = await this.prisma.chatMessage.create({
       data: {
         content: payload.message,
@@ -324,18 +413,7 @@ export class ChatService {
         avatar: client.avatar,
       }
     }
-    console.log("messageWithUser : ", messageWithUser);
-    console.log("message : ", message);
     return messageWithUser;
-  
-    // await this.prisma.roomMembership.update({
-    //   where: {
-    //     id : room.members[0].id,
-    //   },
-    //   data: {
-    //     state: 'BANNED',
-    //   }
-    // })
   }
 
   async getMessages(client: any, payload: any) {
@@ -379,6 +457,11 @@ export class ChatService {
     return messages;
   }
 
+  /**
+  // Small function to help in other functions
+  **/
+
+
   // leave here and move it later to extr file
   async generateUniqueRoomId(length: number = 50){
       const prefix = 'RASY';
@@ -388,6 +471,27 @@ export class ChatService {
       return roomId;
   };
   
+  async isUserInRoom(roomId: number, userId: number): Promise<boolean> {
+    const roomMembership = await this.prisma.roomMembership.findFirst({
+      where: {
+        userId: userId,
+        roomId: roomId,
+      }
+    });
+    console.log('roomMembership : ', !!roomMembership);
+    return !!roomMembership;
+  }
+
+  async isUserAdminInRoom(roomId: number, userId: number): Promise<boolean> {
+    const roomMembership = await this.prisma.roomMembership.findFirst({
+      where: {
+        userId: userId,
+        roomId: roomId,
+        role: 'OWNER' || 'ADMIN',
+      }
+    });
+    return !!roomMembership;
+  }
   // Class END.
   }
